+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed regression in `struct` view when click events aren't handled
+- Added `--discovery-nav-height` as compliment for `--discovery-nav-width`
 
 ## 1.0.0-beta.56 (01-03-2021)